Refactoring contracts on web
