+++ sources/packages/web/src/components/fa-application/EducationProgramList.vue
@@ -16,12 +16,9 @@
 
 <script lang="ts">
 import { SetupContext, ref, onUpdated } from "vue";
-import {
-  InstituteService,
-  EducationProgram,
-  Institute,
-} from "../../services/InstituteService";
-
+import { InstituteService } from "../../services/InstituteService";
+import { EducationProgram } from "../../contracts/EducationContract";
+import { Institute } from "../../contracts/InstituteContract";
 export default {
   emits: ["update:modelValue", "change"],
   props: {
