+++ sources/packages/web/src/contracts/EducationContract.ts
@@ -1,0 +1,4 @@
+export interface EducationProgram {
+  name: string;
+  code?: string;
+}
