+++ sources/packages/web/src/contracts/InstituteContract.ts
@@ -1,0 +1,4 @@
+export interface Institute {
+  name: string;
+  code?: string;
+}
