+++ sources/packages/web/src/contracts/StudentContract.ts
@@ -1,0 +1,51 @@
+export interface Student {
+  phone: string;
+  addressLine1: string;
+  addressLine2: string;
+  city: string;
+  provinceState: string;
+  country: string;
+  postalCode: string;
+}
+
+export interface CreateStudent extends Student {
+  sinNumber: string;
+}
+
+export interface StudentContact {
+  phone: string;
+  addressLine1: string;
+  addressLine2: string;
+  city: string;
+  provinceState: string;
+  country: string;
+  postalCode: string;
+}
+
+export interface StudentInfo {
+  firstName: string;
+  lastName: string;
+  gender: string;
+  email: string;
+  dateOfBirth: Date;
+  contact: {
+    addressLine1: string;
+    addressLine2: string;
+    city: string;
+    provinceState: string;
+    country: string;
+    postalCode: string;
+    phone: string;
+  };
+}
+
+export interface StudentProfile {
+  phone: string;
+  sinNumber: string;
+  addressLine1: string;
+  addressLine2: string;
+  city: string;
+  provinceState: string;
+  country: string;
+  postalCode: string;
+}
