--- sources/packages/web/src/services/http/contracts/StudentContract.ts
@@ -1,16 +1,0 @@
-export interface StudentInfo {
-  firstName: string;
-  lastName: string;
-  gender: string;
-  email: string;
-  dateOfBirth: Date;
-  contact: {
-    addressLine1: string;
-    addressLine2: string;
-    city: string;
-    provinceState: string;
-    country: string;
-    postalCode: string;
-    phone: string;
-  };
-}
